requirements.txt
Flask==3.1.0
langchain_groq==0.2.4
openai==1.72.0
pydantic==2.11.1
pinecone==6.0.1
pymongo==4.11.1
python-dotenv==1.0.1
flask-cors
gunicorn==23.0.0